markdown_to_pdf_gui/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-to-pdf-gui"
version = "1.0.0"
description = "マークダウンファイルをPDFに変換するためのGUI管理ツール"
requires-python = ">=3.8"
dependencies = [
    "PyQt6>=6.6.0",
    "pyyaml>=6.0",
    "cairosvg>=2.7.0",
    "chardet>=5.0.0",
    "pypdf>=3.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-qt>=4.2.0",
]
